requirements.txt
api==0.0.7
appdirs==1.4.3
asgiref==3.2.7
boto3==1.13.17
botocore==1.16.17
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
codegen==1.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-sslserver==0.21
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-firebase==0.3.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
export==0.1.2
filelock==3.0.12
firebase-admin==3.2.0
gcloud==0.17.0
get==2019.4.13
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.29.0
gunicorn==20.0.4
patterns==0.3
Pillow==6.2.1
post==2019.4.13
protobuf==3.10.0
psycopg2==2.8.5
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.4.3
PyJWT==1.7.1
PyMySQL==0.9.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-firebase==1.2
python-jwt==2.0.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==20.0.18
webencodings==0.5.1
whitenoise==5.1.0
